refactor: God bless this mess
