+++ main.cpp
@@ -6,240 +6,15 @@
 #include <Windows.h>
 #include <map>
 
-using namespace std;
-
-class adress {
-public:
-    adress(string city, string district, string street, string num) : city(city), district(district), street(street), num(num) {}
-
-    string getAdress() const {
-        return city + ", " + district + ", " + street + " " + num;
-    }
-
-    string getCity() const { return city; }
-    string getDistrict() const { return district; }
-
-private:
-    string city; // місто
-    string district; // район
-    string street; // вулиця
-    string num; // Номер будинка
-};
-
-class RealEstate { //клас нерухомості
-public:
-    RealEstate(int bud, int floor, double s, bool repair, const adress &X ) : bud(bud), floor(floor), s(s), repair(repair), X(X) {}
-
-    virtual void printInfo() const {
-        cout << "Бюджет: " << bud << "$, Поверх: " << floor  << ", Площа: " << s<< ", Ремонт: " << (repair ? "Так" : "Ні") << ", Адреса: " << X.getAdress() << endl;
-    }
-
-    int getBudget() const { return bud; }
-    double getArea() const { return s; }
-    bool hasRepair() const { return repair; }
-    string getCity() const { return X.getCity(); }
-    string getDistrict() const { return X.getDistrict(); }
-
-    virtual ~RealEstate() {}
-
-protected:
-    int bud; // Бюджет
-    int floor; // Поверхи
-    double s; // Квадратура
-    bool repair; // Ремонт
-    adress X; // Адрес
-};
-
-class House : public RealEstate {
-public:
-    House(int bud, int floor, double s, bool repair, const adress& X, int QualFloor, double TotalArea) : RealEstate(bud, floor, s, repair, X),QualFloor(QualFloor), TotalArea(TotalArea), Dilyanka(true) {}   
-      
-    bool OutSideArea() {   
-        if (s == TotalArea) {
-           return this->Dilyanka = false;
-        }
-    }    
-
-    void printInfo() const override {
-        RealEstate::printInfo();
-        cout << "  [Будинок] Поверхів: " << QualFloor << ", Загальна площа з ділянкою: " << TotalArea << endl;
-    }
-
-private:
-    int QualFloor; // Кількість поверхів
-    double TotalArea; // Розмір ділянки разом з домом
-    bool Dilyanka; // Чи є ділянка в наявності
-};
-
-class Apartment : public RealEstate {
-public:
-    Apartment(int bud, int floor, double s, bool repair, const adress& X, int rooms, bool balcony, int poverh) : RealEstate(bud, floor, s, repair, X), rooms(rooms), balcony(balcony), poverh(poverh){}
-
-    void printInfo() const override {
-        RealEstate::printInfo();
-        cout << "  [Квартира] Кімнат: " << rooms << ", Балкон: " << (balcony ? "Так" : "Ні") << endl;
-    }
-       
-private:
-    int poverh; // на якому поверху знаходиться
-    int rooms;   // кількість кімнат
-    bool balcony; // чи є балкон
-};
-
-class Garage : public RealEstate {
-public:
-    Garage(int bud, int floor, double s, bool repair, const adress& X, bool pit): RealEstate(bud, floor, s, repair, X), pit(pit) {}
-       
-private:
-    bool pit;  // чи є оглядова яма
-};
-
-#include <iostream>
-#include <string>
-#include <vector>
-#include <random>
-#include <ctime>
-#include <Windows.h>
-#include <map>
-#include <algorithm>
-#include <climits>
+#include "./cls/address.hpp"
+#include "./cls/real-estate.hpp"
+#include "./cls/Apartment.hpp"
+#include "./cls/Garage.hpp"
+#include "./cls/house.hpp"
+#include "./cls/Room.hpp"
+#include "./rnd-utils.hpp"
 
 using namespace std;
-
-class adress {
-public:
-    adress(string city, string district, string street, string num) : city(city), district(district), street(street), num(num) {}
-
-    string getAdress() const {
-        return city + ", " + district + ", " + street + " " + num;
-    }
-
-    string getCity() const { return city; }
-    string getDistrict() const { return district; }
-
-private:
-    string city;
-    string district;
-    string street;
-    string num;
-};
-
-class RealEstate {
-public:
-    RealEstate(int bud, int floor, double s, bool repair, const adress& X) : bud(bud), floor(floor), s(s), repair(repair), X(X) {}
-
-    virtual void printInfo() const {
-        cout << "Бюджет: " << bud << "$, Поверх: " << floor << ", Площа: " << s
-            << ", Ремонт: " << (repair ? "Так" : "Ні")
-            << ", Адреса: " << X.getAdress() << endl;
-    }
-
-    int getBudget() const { return bud; }
-    double getArea() const { return s; }
-    bool hasRepair() const { return repair; }
-    string getCity() const { return X.getCity(); }
-    string getDistrict() const { return X.getDistrict(); }
-
-    virtual ~RealEstate() {}
-
-protected:
-    int bud;
-    int floor;
-    double s;
-    bool repair;
-    adress X;
-};
-
-class House : public RealEstate {
-public:
-    House(int bud, int floor, double s, bool repair, const adress& X, int QualFloor, double TotalArea)
-        : RealEstate(bud, floor, s, repair, X), QualFloor(QualFloor), TotalArea(TotalArea), Dilyanka(true) {
-    }
-
-    void printInfo() const override {
-        RealEstate::printInfo();
-        cout << "  [Будинок] Поверхів: " << QualFloor
-            << ", Загальна площа з ділянкою: " << TotalArea << endl;
-    }
-
-private:
-    int QualFloor;
-    double TotalArea;
-    bool Dilyanka;
-};
-
-class Apartment : public RealEstate {
-public:
-    Apartment(int bud, int floor, double s, bool repair, const adress& X, int rooms, bool balcony, int poverh)
-        : RealEstate(bud, floor, s, repair, X), rooms(rooms), balcony(balcony), poverh(poverh) {
-    }
-
-    void printInfo() const override {
-        RealEstate::printInfo();
-        cout << "  [Квартира] Кімнат: " << rooms
-            << ", Балкон: " << (balcony ? "Так" : "Ні") << endl;
-    }
-
-    int getRooms() const { return rooms; }
-    bool hasBalcony() const { return balcony; }
-
-private:
-    int poverh;
-    int rooms;
-    bool balcony;
-};
-
-class Garage : public RealEstate {
-public:
-    Garage(int bud, int floor, double s, bool repair, const adress& X, bool pit)
-        : RealEstate(bud, floor, s, repair, X), pit(pit) {
-    }
-
-    bool hasPit() const { return pit; }
-
-    void printInfo() const override {
-        RealEstate::printInfo();
-        cout << "  [Гараж] Оглядова яма: " << (pit ? "Так" : "Ні") << endl;
-    }
-
-private:
-    bool pit;
-};
-
-class Room : public RealEstate {
-public:
-    Room(int bud, int floor, double s, bool repair, const adress& X, bool kitchen, bool bath)
-        : RealEstate(bud, floor, s, repair, X), kitchen(kitchen), bath(bath) {
-    }
-
-    bool hasKitchen() const { return kitchen; }
-    bool hasBath() const { return bath; }
-
-    void printInfo() const override {
-        RealEstate::printInfo();
-        cout << "  [Кімната] Кухня: " << (kitchen ? "Так" : "Ні")
-            << ", Ванна: " << (bath ? "Так" : "Ні") << endl;
-    }
-
-private:
-    bool kitchen;
-    bool bath;
-};
-
-// ==== Генератор випадкових чисел ====
-int rnd(int a, int b) {
-    static random_device rd;
-    static mt19937 gen(rd());
-    uniform_int_distribution<> dist(a, b);
-    return dist(gen);
-}
-
-double rndDouble(double a, double b) {
-    static random_device rd;
-    static mt19937 gen(rd());
-    uniform_real_distribution<> dist(a, b);
-    return dist(gen);
-}
 
 // ==== Генерація даних ====
 vector<RealEstate*> generateRealEstate() {
